react-hot-loader and css dev updates
