+++ .babelrc
@@ -1,6 +1,7 @@
 {
   "presets": ["react", "es2015"],
   "plugins": [
+    "react-hot-loader/babel",
     "transform-object-rest-spread",
     [
       "module-resolver",
