+++ dev/index.html
@@ -4,30 +4,6 @@
     <meta charset="utf-8">
     <link rel="shortcut icon" href="/favicon.ico">
     <title>Editor Dev</title>
-    <style>
-      body {
-        margin: 0;
-        padding: 0;
-        font-family: sans-serif;
-      }
-
-      .app{
-        display: flex;
-        /*
-           We are defining the max height of the app so that the editor knows how big to be
-           currently the editor will take up whatever space it can if it is not constrained in its parent
-         */
-        min-height: 100vh;
-        max-height: 100vh;
-        width: 100%;
-      }
-      .app__main {
-        max-width: 100%;
-        max-height: 100vh;
-        overflow: auto;
-        flex-grow: 1;
-      }
-    </style>
   </head>
   <body>
     <div id="root"></div>
