+++ package.json
@@ -1,45 +1,28 @@
 {
   "name": "react-plotly.js-editor",
-  "version": "0.7.0",
-  "license": "MIT",
   "description": "plotly.js chart editor react component UI",
+  "version": "0.7.0",
   "author": "Plotly, Inc.",
-  "main": "lib/index.js",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/plotly/react-plotly.js-editor.git"
-  },
   "bugs": {
     "url": "https://github.com/plotly/react-plotly.js-editor/issues"
   },
-  "scripts": {
-    "start": "webpack-dev-server",
-    "lint": "prettier --write \"src/**/*.js\"",
-    "make:combined-translation-keys": "npm run make:translation-keys && node scripts/combineTranslationKeys.js",
-    "make:translation-keys": "node scripts/findTranslationKeys.js",
-    "make:arrows": "node scripts/makeArrows.js",
-    "make:lib": "mkdirp lib && npm run make:lib:js && npm run make:lib:css && npm run make:combined-translation-keys",
-    "make:lib:js": "mkdirp lib && babel src --out-dir lib --ignore=__tests__/* --source-maps",
-    "make:lib:css": "mkdirp lib && babel-node scripts/styles.js && SASS_ENV=ie babel-node scripts/styles.js && babel-node scripts/postcss.js && SASS_ENV=ie babel-node scripts/postcss.js",
-    "prepublishOnly": "npm run make:lib",
-    "test:lint": "eslint \"src/**/*.js\" && echo -e '\\033[0;32m'PASS'\\033[0m'",
-    "test:pretty": "prettier -l \"src/**/*.js\" && echo -e '\\033[0;32m'PASS'\\033[0m'",
-    "test:js": "jest --setupTestFrameworkScriptFile=raf/polyfill",
-    "test": "npm run test:lint && npm run test:pretty && npm run test:js",
-    "watch": "babel src --watch --out-dir lib --source-maps | node-sass -w src/styles/main.scss lib/react-plotly.js-editor.css",
-    "watch-test": "jest --watch"
+  "dependencies": {
+    "draft-js": "^0.10.4",
+    "draft-js-export-html": "github:plotly/draft-js-export-html",
+    "draft-js-import-html": "^1.2.1",
+    "draft-js-utils": "^1.2.0",
+    "fast-isnumeric": "^1.1.1",
+    "plotly-icons": "^1.1.3",
+    "plotly.js": "^1.33.0",
+    "prop-types": "^15.5.10",
+    "raf": "^3.4.0",
+    "react": "^16.2.0",
+    "react-color": "^2.13.8",
+    "react-colorscales": "^0.4.2",
+    "react-dom": "^16.2.0",
+    "react-select": "^1.0.0-rc.10",
+    "tinycolor2": "^1.4.1"
   },
-  "keywords": [
-    "graphing",
-    "plotting",
-    "data",
-    "visualization",
-    "plotly",
-    "react",
-    "editor",
-    "chart",
-    "widgets"
-  ],
   "devDependencies": {
     "autoprefixer": "^7.2.3",
     "babel-cli": "^6.26.0",
@@ -76,6 +59,7 @@
     "postcss-custom-properties": "^6.2.0",
     "postcss-remove-root": "^0.0.2",
     "prettier": "^1.9.2",
+    "react-hot-loader": "^4.0.0-beta.21",
     "react-plotly.js": "^1.2.0",
     "react-test-renderer": "^16.2.0",
     "sass-loader": "^6.0.6",
@@ -83,27 +67,7 @@
     "webpack": "^3.10.0",
     "webpack-dev-server": "^2.11.1"
   },
-  "peerDependencies": {
-    "react": "^15.3.0 || ^16.0.0-rc || ^16.0",
-    "react-dom": "^0.14.9 || ^15.3.0 || ^16.0.0-rc || ^16.0"
-  },
-  "dependencies": {
-    "draft-js": "^0.10.4",
-    "draft-js-export-html": "github:plotly/draft-js-export-html",
-    "draft-js-import-html": "^1.2.1",
-    "draft-js-utils": "^1.2.0",
-    "fast-isnumeric": "^1.1.1",
-    "plotly-icons": "^1.1.3",
-    "plotly.js": "^1.33.0",
-    "prop-types": "^15.5.10",
-    "raf": "^3.4.0",
-    "react": "^16.2.0",
-    "react-color": "^2.13.8",
-    "react-colorscales": "^0.4.2",
-    "react-dom": "^16.2.0",
-    "react-select": "^1.0.0-rc.10",
-    "tinycolor2": "^1.4.1"
-  },
+  "homepage": "https://plotly.github.io/react-plotly.js-editor/",
   "jest": {
     "roots": [
       "<rootDir>/src/"
@@ -111,5 +75,43 @@
     "moduleNameMapper": {
       "^.+\\.css$": "<rootDir>/config/CSSStub.js"
     }
+  },
+  "keywords": [
+    "chart",
+    "data",
+    "editor",
+    "graphing",
+    "plotly",
+    "plotting",
+    "react",
+    "visualization",
+    "widgets"
+  ],
+  "license": "MIT",
+  "main": "lib/index.js",
+  "peerDependencies": {
+    "react": "^15.3.0 || ^16.0.0-rc || ^16.0",
+    "react-dom": "^0.14.9 || ^15.3.0 || ^16.0.0-rc || ^16.0"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/plotly/react-plotly.js-editor.git"
+  },
+  "scripts": {
+    "lint": "prettier --write \"src/**/*.js\"",
+    "make:arrows": "node scripts/makeArrows.js",
+    "make:combined-translation-keys": "npm run make:translation-keys && node scripts/combineTranslationKeys.js",
+    "make:lib": "mkdirp lib && npm run make:lib:js && npm run make:lib:css && npm run make:combined-translation-keys",
+    "make:lib:css": "mkdirp lib && babel-node scripts/styles.js && SASS_ENV=ie babel-node scripts/styles.js && babel-node scripts/postcss.js && SASS_ENV=ie babel-node scripts/postcss.js",
+    "make:lib:js": "mkdirp lib && babel src --out-dir lib --ignore=__tests__/* --source-maps",
+    "make:translation-keys": "node scripts/findTranslationKeys.js",
+    "prepublishOnly": "npm run make:lib",
+    "start": "webpack-dev-server --hot",
+    "test": "npm run test:lint && npm run test:pretty && npm run test:js",
+    "test:js": "jest --setupTestFrameworkScriptFile=raf/polyfill",
+    "test:lint": "eslint \"src/**/*.js\" && echo -e '\\033[0;32m'PASS'\\033[0m'",
+    "test:pretty": "prettier -l \"src/**/*.js\" && echo -e '\\033[0;32m'PASS'\\033[0m'",
+    "watch": "babel src --watch --out-dir lib --source-maps | node-sass -w src/styles/main.scss lib/react-plotly.js-editor.css",
+    "watch-test": "jest --watch"
   }
 }
